requirements.txt
acachecontrol==0.3.5
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
decorator==5.1.1
ecdsa==0.17.0
Flask==2.0.3
frozenlist==1.3.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
multidict==6.0.2
okta-jwt-verifier==0.2.3
py==1.11.0
pyasn1==0.4.8
PyJWT==2.3.0
python-jose==3.3.0
requests==2.27.1
retry==0.9.2
rsa==4.8
six==1.16.0
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.0.3
yarl==1.7.2
sqlalchemy
psycopg2-binary